+++ examples/Example.ParseAndBind.cat/Properties/AssemblyInfo.cs
@@ -1,21 +1,15 @@
 ﻿
-// Created: 10th August 2009
-// Updated: 8th June 2019
-
 #pragma warning disable 1607
 
 using System;
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("CLASP.Util")]
-[assembly: AssemblyDescription("CLASP.NET utility API")]
+[assembly: AssemblyTitle("Example.ParseAndBind.cat")]
+[assembly: AssemblyDescription("")]
 
 [assembly: CLSCompliant(true)]
 [assembly: ComVisible(false)]
-
-[assembly: AssemblyVersion("0.22.153.1")]
-[assembly: AssemblyFileVersion("0.22.153.1")]
 
 /* ///////////////////////////// end of file //////////////////////////// */